Complete ComWrappersObject integration and coverage
Add Mono availability, pointer-sized holder IDs with native layout validation, and field-backed RCW lookup in DAC and cDAC. Preserve tracker publication ordering and exact-handle cache lifetime. Streamline the known IUnknown return path and extend identity, concurrency, resurrection, and diagnostic coverage.

Copilot-Session: 2a77edfd-ecc2-4e25-91b5-715ac3cb5fca

+++ src/coreclr/vm/interoplibinterface_comwrappers.h
@@ -79,7 +79,7 @@ public:
     DPTR(InteropLib::ABI::ManagedObjectWrapperLayout) _wrapper;
     // Identifies the ComWrappers instance that created this wrapper. Unused natively, but the managed and
     // native declarations of this type must agree on size, which CoreLibBinder verifies on startup.
-    uint64_t _comWrappersId;
+    uintptr_t _comWrappersId;
     friend struct ::cdac_data<ManagedObjectWrapperHolderObject>;
 };
